changed a tag to button tag
